+++ GUI/Controls/ChooseRecommendedMods.cs
@@ -26,15 +26,15 @@ namespace CKAN.GUI
         }
 
         [ForbidGUICalls]
-        public void LoadRecommendations(IRegistryQuerier    registry,
-                                        List<CkanModule>    toInstall,
-                                        HashSet<CkanModule> toUninstall,
-                                        GameVersionCriteria versionCrit,
-                                        NetModuleCache      cache,
-                                        IGame               game,
-                                        List<ModuleLabel>   labels,
-                                        IConfiguration      coreConfig,
-                                        GUIConfiguration    guiConfig,
+        public void LoadRecommendations(IRegistryQuerier        registry,
+                                        ICollection<CkanModule> toInstall,
+                                        ICollection<CkanModule> toUninstall,
+                                        GameVersionCriteria     versionCrit,
+                                        NetModuleCache          cache,
+                                        IGame                   game,
+                                        List<ModuleLabel>       labels,
+                                        IConfiguration          coreConfig,
+                                        GUIConfiguration        guiConfig,
                                         Dictionary<CkanModule, Tuple<bool, List<string>>> recommendations,
                                         Dictionary<CkanModule, List<string>>              suggestions,
                                         Dictionary<CkanModule, HashSet<string>>           supporters)
@@ -269,6 +269,11 @@ namespace CKAN.GUI
             CheckUncheckRows(RecommendationsGroup.Items, true);
         }
 
+        private void CheckSuggestionsButton_Click(object? sender, EventArgs? e)
+        {
+            CheckUncheckRows(SuggestionsGroup.Items, true);
+        }
+
         private void CheckUncheckRows(ListView.ListViewItemCollection items,
                                       bool check)
         {
@@ -320,12 +325,12 @@ namespace CKAN.GUI
             RecommendedModsListView.ItemChecked -= RecommendedModsListView_ItemChecked;
         }
 
-        private IRegistryQuerier?    registry;
-        private List<CkanModule>     toInstall = new List<CkanModule>();
-        private HashSet<CkanModule>  toUninstall = new HashSet<CkanModule>();
-        private GameVersionCriteria? versionCrit;
-        private GUIConfiguration?    guiConfig;
-        private IGame?               game;
+        private IRegistryQuerier?       registry;
+        private ICollection<CkanModule> toInstall   = Array.Empty<CkanModule>();
+        private ICollection<CkanModule> toUninstall = Array.Empty<CkanModule>();
+        private GameVersionCriteria?    versionCrit;
+        private GUIConfiguration?       guiConfig;
+        private IGame?                  game;
         private TaskCompletionSource<HashSet<CkanModule>?>? task;
     }
 }