+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - [Multiple] Make global install filters and presets game-specific (#4361 by: HebaruSan)
 - [GUI] Add download count to Metadata Tab (#4363, #4368 by: averageksp; reviewed: HebaruSan)
 - [GUI] Calculate size of cached files per host (#4377 by: HebaruSan)
+- [GUI] Show recommendations of chosen recommendations in GUI, readable foreground colors for labels (#4381 by: HebaruSan)
 
 ### Bugfixes